+++ polybot/tests/test_planner.py
@@ -1,6 +1,7 @@
 """Make sure the planning system works"""
 import json
 import logging
+import time
 from time import sleep
 
 from colmena.models import Result
@@ -34,22 +35,18 @@ def test_get_sample_from_url(mocker: MockerFixture):
     assert isinstance(opt.search_template, SampleTemplate)
 
 
-def test_generate(mocker: MockerFixture, opt_config, example_sample, caplog):
+def test_generate(mocker: MockerFixture, mock_subscribe, opt_config, example_sample, caplog):
     # Mock the send_new_sample in the planning library
     fake_robot = mocker.patch('polybot.planning.send_new_sample')
 
     # Make the planner
     client_q = settings.make_client_queue()
-    server_q = settings.make_server_queue()
     planner = RandomPlanner(client_q, opt_config, daemon=True)
 
     # Launch it as a Thread
     planner.start()
+    time.sleep(5)  # Wait for the thread to start up and the subscription to download a sample
     try:
-        # Test sending in a new result directly via Redis
-        server_q.send_result(Result(((0,), {})), topic='robot')
-        sleep(1)  # For the other thread to catch up
-
         # Make sure the server responded by sending a record to the "robot"
         assert fake_robot.call_count == 1, planner.is_alive()
     finally:
